requirements.txt
fastapi
uvicorn
prometheus-fastapi-instrumentator
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
grpcio 
grpcio-tools
prometheus_client
